refactor: 매년 반복 옵션 컴포넌트 분리
